backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0
openai==1.10.0

# Testing
pytest==8.0.0
pytest-cov==4.1.0
httpx==0.27.0
starlette[full]==0.35.1

# Linting
ruff==0.1.14
